refactor: separate out FileLock as a common module (ai.backend.storage.filelock)
